requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==22.0.0
asyncpg==0.29.0
httpx==0.27.2
aiogram==3.13.1
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
pyjwt==2.9.0
jinja2==3.1.4
python-dotenv==1.0.1
openai==1.47.0
rich==13.8.1
python-multipart==0.0.9